# Various improvements to performance and stability