--- a/v2/go.mod
+++ b/v2/go.mod
@@ -1,8 +1,14 @@
 module gomodules.xyz/jsonpatch/v2
 
-go 1.12
+go 1.20
 
 require (
 	github.com/evanphx/json-patch v0.5.2
 	github.com/stretchr/testify v1.3.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 )